Move Histogram and Counter into collector module
