+++ lib/phobos_prometheus.rb
@@ -9,22 +9,24 @@ require 'prometheus/client/formats/text'
 require 'sinatra/base'
 
 require 'phobos_prometheus/version'
-require 'phobos_prometheus/histogram_collector'
-require 'phobos_prometheus/counter_collector'
+require 'phobos_prometheus/collector'
+require 'phobos_prometheus/collector/histogram'
+require 'phobos_prometheus/collector/counter'
 require 'phobos_prometheus/exporter_helper'
 require 'phobos_prometheus/exporter'
 
 # Prometheus collector for Phobos
 module PhobosPrometheus
   class << self
-    attr_reader :config
+    attr_reader :config, :metrics
 
     def subscribe
       config.metrics.each do |metric|
         metric.types.each do |type|
-          register(type: type,
-                   instrumentation_label: metric.instrumentation_label,
-                   buckets: bucket_config(metric.bucket))
+          metrics << register(
+            type: type, buckets: bucket_config(metric.bucket),
+            instrumentation_label: metric.instrumentation_label
+          )
         end
       end
 
@@ -34,6 +36,7 @@ module PhobosPrometheus
     end
 
     def configure(configuration)
+      @metrics ||= []
       @config = Phobos::DeepStruct.new(fetch_settings(configuration))
 
       Phobos.logger.info { Hash(message: 'PhobosPrometheus configured', env: ENV['RACK_ENV']) }
@@ -50,9 +53,9 @@ module PhobosPrometheus
     def register(type:, instrumentation_label:, buckets:)
       case type
       when 'counter'
-        CounterCollector.create(instrumentation_label: instrumentation_label)
+        Collector::Counter.create(instrumentation_label: instrumentation_label)
       when 'histogram'
-        PhobosPrometheus::HistogramCollector.create(
+        Collector::Histogram.create(
           instrumentation_label: instrumentation_label,
           buckets: buckets
         )
