+++ lib/phobos_prometheus/collector.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module PhobosPrometheus
+  module Collector
+  end
+end
